+++ src/Spice86/Emulator/LoadableFile/Dos/Com/ComLoader.cs
@@ -6,17 +6,17 @@ using Spice86.Emulator.Memory;
 using Spice86.Emulator.CPU;
 
 public class ComLoader : ExecutableFileLoader {
-    private const int ComOffset = 0x100;
-    private readonly int _startSegment;
+    private const ushort ComOffset = 0x100;
+    private readonly ushort _startSegment;
 
-    public ComLoader(Machine machine, int startSegment) : base(machine) {
+    public ComLoader(Machine machine, ushort startSegment) : base(machine) {
         this._startSegment = startSegment;
     }
 
     public override byte[] LoadFile(string file, string? arguments) {
         new PspGenerator(_machine).GeneratePsp(_startSegment, arguments);
         byte[] com = this.ReadFile(file);
-        int physicalStartAddress = MemoryUtils.ToPhysicalAddress(_startSegment, ComOffset);
+        uint physicalStartAddress = MemoryUtils.ToPhysicalAddress(_startSegment, ComOffset);
         _memory.LoadData(physicalStartAddress, com);
         State? state = _cpu.GetState();
